Move storeutil package to example/todomvc/store.
Resolves #137.
