+++ example/todomvc/store/storeutil/storeutil.go
